# packages/contracts/openapi/depsis.yaml
openapi: 3.1.0
info:
  title: DEPSIS API
  version: 0.1.0-phase1
  summary: Debian tabanlı çok kullanıcılı NAS platformu — Faz 1 yüzeyi
  description: |
    Bu belge tek doğruluk kaynağıdır. `apps/web` istemcisi buradan üretilir
    (`openapi-typescript`, ADR-0002); elle yazılmış istemci tipi bulunamaz.

    Kapsam: Faz 1 uçları. Konsol, uygulama kataloğu ve uzak erişim de burada — sırasıyla ADR-0018, ADR-0019 ve ADR-0020. Üçü de cihazın DIŞINDAKİ bir şeyi yönetiyor (bir kabuk, podman, zerotier-one); DEPSIS hiçbirini paketlemez, kuruluysa yönetir ve kurulu değilse 503 döner. Konsol ayrıcalıklı ajanda ÇALIŞMAZ: §2.2'nin serbest komut yasağı duruyor.
    Faz 2-3'e aittir ve kasıtlı olarak yoktur — tanımlanmamış bir uç, "yakında" diye
    duran boş bir uçtan iyidir.

    Sistem aracısı IPC şeması burada DEĞİLDİR. ADR-0006 gereği onu Rust sahiplenir ve
    schemars ile üretir: bir güven sınırının sözleşmesini, o sınırı uygulayan taraf
    tanımlamalıdır.
servers:
  - url: /api/v1
...
security:
  - sessionCookie: []
paths:
...
  /files:
    get:
      tags: [files]
      summary: Bir klasörün içeriğini listele
      description: |
        Cursor pagination zorunlu (§14). Offset kullanılmaz: bir NAS'ta küme sürekli
        değiştiği için offset sayfalama sessizce kayıt atlar.

        Yanıtta toplam sayı yoktur, yalnız hasMore. Filtrelenmiş bir liste yanında
        filtrelenmemiş bir sayaç, kiracının göremediği satırların varlığını sızdırır.
      parameters:
        - name: parentId
          in: query
          description: Yoksa kullanıcının kökleri döner.
          schema: { type: string, format: uuid }
        - name: shareId
          in: query
          required: false
          description: |
            Hangi paylaşım. Verilmezse kiracının varsayılanı — yani bu parametreden önceki her
            istemci aynen çalışmaya devam eder.

            Bu parametre olmadan dosya yöneticisi HER ZAMAN ilk paylaşımı gösteriyordu
            (`ORDER BY created_at LIMIT 1`), yani `POST /shares` ile açılan bir paylaşım
            Samba'ya yayımlanıyor ve web'de görünmüyordu.

            Başka bir kiracının paylaşımı 404 — var olmayan bir kimlikle aynı cevap, yoksa
            parametre hangi paylaşım kimliklerinin var olduğunu söyleyen bir kâhin olurdu.
          schema: { type: string, format: uuid }
        - name: trashed
          in: query
          description: |
            Doğruysa çöp kutusu listelenir ve `parentId` yok sayılır. Çöp bir klasör değil bir
            sütun (0008: `trashed_at`), yani ağaçta gezilecek bir yeri yok; ayrı bir uç nokta
            yerine bir süzgeç olması da bu yüzden.
          schema: { type: boolean, default: false }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, modified, size]
            default: name
      responses:
        '200':
          description: Sayfa
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileEntryPage' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
...
  /files/{id}:
    parameters:
      - $ref: '#/components/parameters/FileId'
    get:
      tags: [files]
      summary: Tek bir girdinin ayrıntısı
      responses:
        '200':
          description: Girdi
          headers:
            ETag:
              description: |
                Meta veri doğrulayıcısı. `PATCH /files/{id}` bunu `If-Match` olarak kabul eder.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileEntryDetail' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
    patch:
      tags: [files]
      summary: Yeniden adlandır
      description: |
        Taşıma da burada: `parentId` alanı aynı satırın diğer kolonu ve dosya sistemi tarafında da tek bir rename(2). PAYLAŞIMLAR ARASI taşıma burada değil — dataset sınırını geçen bir rename EXDEV veriyor (ADR-0008) ve kopyala+sil işine dönüşmesi gerek; o /file-operations — taşımada hem kaynak
        hem hedef yetkisi gerekir (§6.2).
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateFileRequest' }
      responses:
        '200':
          description: Güncellendi
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileEntry' }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '412': { $ref: '#/components/responses/Problem' }
        # 501 KALDIRILDI. Taşıma artık ajanın `MoveEntry` işlemiyle gerçekten yapılıyor:
        # önce dosya sistemi, sonra veritabanı. Sıra bu yönde çünkü ters sırada, rename
        # başarısız olduğunda satır yeni yerini gösterir ve diskteki dosya eski yerinde kalır —
        # veritabanıyla diskin ayrışması, bu üründe kabul edilmeyen tek şey.
    delete:
      tags: [files]
      summary: Çöp kutusuna taşı
      description: |
        Bu bir JOB DEĞİL. Önceki hâli 202 + JobRef diyordu, çünkü büyük bir alt ağacı silmek
        uzun iş gibi görünüyordu; çöpe atmak tek satırda tek bayrak çıktı. Bayt silinmiyor —
        çöpü boşaltmak ajandan `unlink` isteyen ayrı bir karar. Var olmayan bir işe referans
        döndürmek, istemcinin sorgulayamayacağı bir kuyruk girdisi uydurmak olurdu.
      responses:
        '200':
          description: Çöp kutusuna taşındı
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileEntry' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
...
  /file-operations:
    post:
      tags: [files]
      summary: Kopyala — uzun süren iş olarak
      description: |
        Hem kaynak hem hedef yetkisi doğrulanır (§6.2): kaynakta `download`, hedefte `create`.
        `download`, `read` değil — `read` üstveridir (ad, boyut, tarih), kopyalama İÇERİĞİ alır.
        Bir dosyanın var olduğunu görebilmek, onun kopyasını kendi klasörüne çıkarabilmek
        anlamına gelmemeli.

        202 ve bir `JobRef` dönüyor. §17: bin dosyalık toplu bir işlem arayüzü kilitlemez ve
        durumu gerçek zamanlı güncellenir — iş kuyrukta çalışıyor, `GET /jobs/{jobId}` ve
        `GET /events` üzerinden izleniyor.

        ── Şu an ne sunuluyor ──────────────────────────────────────────────────

        `operation: copy` — evet.

        `operation: move` — hayır, 422. Tek girdi taşıma `PATCH /files/{id}`: tek bir
        `renameat2`, anında, ve kuyruğa sokmak yaygın durumu hiçbir kazanç olmadan yavaşlatır.
        PAYLAŞIMLAR ARASI taşıma gerçekten buraya ait — dataset sınırını geçen bir rename EXDEV
        verir (ADR-0008) ve kopyala+sil'e dönüşmesi gerekir — ama yarım uygulanmış bir taşıma,
        kopyalayıp silmeyi başaramadığında kullanıcının bir tane istediği yerde iki dosya
        bırakır. Yaklaşık yapmak yerine reddediliyor.

        `operation: restore` — hayır, 422. `POST /files/{id}/restore` var.

        `conflictPolicy: keep_both` — evet, ve şu an tek desteklenen. `replace` kullanıcının
        adlandırmadığı bir dosyayı yok ederdi; `RENAME_NOREPLACE` bunu syscall'a kadar
        engelliyor ve desteklemek ajana sahip olmadığı bir üzerine-yazma vermek demek olurdu.
        `version` var olmayan bir sürüm deposu istiyor. `skip` savunulabilir ve henüz yazılmadı;
        onun adı altında sessizce `keep_both` yapmaktansa reddetmek daha dürüst.

        ── Sınırlar ────────────────────────────────────────────────────────────

        Bir klasör kendi içine ya da kendi altındaki bir klasöre kopyalanamaz: 409. Her adım tek
        başına geçerli bir oluşturma olduğu için veritabanı bunu durdurmaz, sonu gelmeyen bir
        ağaç olurdu.

        100.000 girdiden büyük bir seçim reddediliyor. Performans sınırı değil: kopyalama, tek
        bir yükleme olmadan saklanan baytları çoğaltabilen iki işlemden biri, ve paylaşımın
        kökünü seçen birine sayıyı söylemek, bir saat sonra dolmuş bir dataset olarak
        keşfettirmekten iyi.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FileOperationRequest' }
      responses:
        '202':
          description: İş kuyruğa alındı
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobRef' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
        '507':
          description: |
            Seçimin toplam boyutu havuzda kalan yerden büyük. Bir garanti değil bir nezaket
            kontrolü — eşzamanlılık altında yer, bu cevapla kopyalama arasında başka bir yükleme
            tarafından alınabilir; ajanın ENOSPC/EDQUOT sınıflandırması bu yüzden ayrıca var.
            Dönüştürdüğü şey yaygın durum: bir saat sonra yarısı kopyalanmış bir ağaçla
            başarısız olmak yerine, iki sayıyı hemen söylemek.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
...
  /tasks/{id}/files:
    get:
      tags: [tasks]
      summary: Göreve bağlı dosyalar
      description: |
        **Görev erişimi dosya erişimi vermiyor** (§7). Bu uç her bağ için dosyanın KENDİ izinlerini
        çözüyor ve çağıranın göremeyeceği satırları hiç döndürmüyor — göremediği bir dosyanın adı,
        yolu ya da paylaşımı bu yanıtta geçmiyor.

        Görülemeyenlerin SAYISI `hidden` alanında. Sıfır göstermek "burada başka bir şey yok"
        demek olurdu ve bu yanlış; sayıyı tamamen gizlemek bir görevi eksik gösterirdi. Sayı,
        "buraya bakman gereken bir şey var, ama sana değil" cümlesinin en az bilgi sızdıran hâli.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Bağlar
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskFileLinkPage' }
        '404': { $ref: '#/components/responses/Problem' }
    post:
      tags: [tasks]
      summary: Göreve dosya bağla
      description: |
        Bağlayanın o dosyada **en az `read`** izni olmak zorunda. Göremediği bir dosyayı bağlamak,
        onu görebilen birine "bu dosya var" demenin dolaylı yolu olurdu — ve daha kötüsü, bağın
        kendisi bir izin gibi okunabilirdi.

        Göremediği bir dosya için cevap **404**, 403 değil: 403 dosyanın var olduğunu söyler.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LinkTaskFileRequest' }
      responses:
        '201':
          description: Bağlandı
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskFileLink' }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /tasks/{id}/files/{linkId}:
    delete:
      tags: [tasks]
      summary: Bağı kaldır
      description: |
        Dosyaya dokunmuyor — yalnız bağı siliyor. Bağı kaldırmak için dosyayı GÖREBİLMEK gerekmiyor:
        göremediğin bir dosyanın bağı listende duruyorsa, onu kaldırabilmen gerekir, yoksa
        temizleyemeyeceğin bir satır kalır.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: linkId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Kaldırıldı }
        '404': { $ref: '#/components/responses/Problem' }
...
  /shares/{shareId}/snapshots:
    get:
      operationId: listShareSnapshots
      tags: [files]
      summary: Bu paylaşımın geçmiş sürümleri
      description: |
        Paylaşımın veri kümesindeki anlık görüntüler, yenisi önce. HAVUZDAN okunuyor, DEPSIS'in
        kendi kaydından değil: kabuktan silinmiş bir görüntü kayıtta durur, ve ekranın var olmayan
        bir geri dönüş noktası önermesi, okuyanın bunu tam ihtiyaç duyduğu an öğrenmesi demektir.

        `available: false` ajana ulaşılamadığında. Boş liste DEĞİL: boş liste "hiç görüntü yok"
        demek olurdu, ve bir dakikalığına düşmüş bir ajan bütün yedekleri silinmiş gibi gösterirdi.

        YETKİ: paylaşımın KÖKÜNDE `download`. Bir anlık görüntüde gezinmek, o paylaşımın geçmişteki
        bütün adlarını görmek demek, ve klasör bazlı bir izin bu ağaca eşlenemiyor — geri
        yüklenecek klasör çoğu zaman artık yok. Dar taraf seçildi: yalnız alt klasörüne yetkisi
        olan biri geçmiş sürümlere hiç bakamıyor.
      parameters:
        - name: shareId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Anlık görüntüler
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ShareSnapshotPage' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /shares/{shareId}/snapshots/{snapshot}/entries:
    get:
      operationId: listSnapshotEntries
      tags: [files]
      summary: Bir anlık görüntünün içinde bir klasör
      description: |
        Anlık görüntünün içindeki TEK bir klasörün içeriği. Ağaç değil — çağıranın maliyetini
        seçtiği bir çağrı olmaması için (ADR-0006 §2.2), arayüz ağacı kendisi geziyor.

        Salt okunur, ve başka türlüsü mümkün değil: bir ZFS anlık görüntüsü değiştirilemez.
      parameters:
        - name: shareId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: snapshot
          in: path
          required: true
          description: Görüntünün kendi adı — `@`'den sonraki kısım.
          schema: { type: string }
        - name: path
          in: query
          required: false
          description: |
            Görüntünün kökünden itibaren, `/` ile ayrılmış. Boş bırakılırsa paylaşımın kökünün o
            günkü hâli.
          schema: { type: string }
      responses:
        '200':
          description: Klasörün içeriği
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SnapshotListing' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /shares/{shareId}/snapshots/{snapshot}/restore:
    post:
      operationId: restoreFromSnapshot
      tags: [files]
      summary: Bir dosyayı anlık görüntüden geri getir
      description: |
        Anlık görüntüdeki BİR dosyayı canlı paylaşıma kopyalar. Veri kümesini geri almaz — bunu
        yapmak, o görüntüden beri yazılmış her şeyi de atmak olurdu.

        ÜZERİNE YAZMAZ. Hedefteki ad doluysa sunucu boş bir ad seçiyor ve SEÇTİĞİ ADI cevapta
        döndürüyor, böylece kullanıcı dosyanın nereye ineceğini onaylamadan önce görüyor. Ajan
        tarafında da `RENAME_NOREPLACE` var: iki kapı, çünkü geri yüklemenin tamamı "hangi kopyayı
        istediğimden emin değilim" demek.

        Bir İŞ döndürüyor. Ajanın denetim soketi aynı anda tek bağlantıya hizmet ediyor, yani
        baytlar dilimler hâlinde taşınıyor ve büyük bir dosya hiçbir HTTP isteğinin beklememesi
        gereken kadar sürüyor.

        YETKİ: paylaşımın kökünde `download`, hedef klasörde `create`.
      parameters:
        - name: shareId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: snapshot
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RestoreRequest' }
      responses:
        '202':
          description: Geri yükleme kuyruğa alındı
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RestoreAccepted' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
...
  /remote/controller:
    get:
      operationId: getControllerStatus
      tags: [remote]
      summary: Bu cihaz kendi ağını yönetebilir mi
      description: |
        `zerotier-one`'ın kendisi controller — ayrı bir servis yok. Bu uç, daemon'un ayakta olup
        olmadığını ve cihazın kendi 10 haneli adresini söylüyor.

        ADRES SADECE GÖSTERİM İÇİN DEĞİL: üye listesinde cihazın kendi satırı, "yetkiyi kaldır"
        düğmesi ASLA gösterilmemesi gereken tek satır. Kendi yetkisini kaldıran bir cihaz, kendi
        sunduğu ağdan düşüyor ve geri almanın yolu tam da kopan bağlantının arkasında kalıyor.

        Yalnız yönetici.
      responses:
        '200':
          description: Controller durumu
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ControllerStatus' }
        '403': { $ref: '#/components/responses/Problem' }
        '503': { $ref: '#/components/responses/Problem' }
...
  /files/{id}/restore:
    parameters:
      - $ref: '#/components/parameters/FileId'
    post:
      tags: [files]
      summary: Çöp kutusundan geri al
      description: |
        Geri alma, adın hâlâ boş olmasını gerektirir: çöpe atılmış bir dosya adını rehin tutmaz
        (0008'deki kısmi benzersizlik indeksleri), yani aynı adla yenisi oluşturulmuş olabilir.
        Böyle bir durumda 409 döner ve kullanıcıdan yeniden adlandırması beklenir — sessizce
        "(1)" eklemek, iki dosyadan hangisinin hangisi olduğunu kimsenin bilemeyeceği yerdir.
      responses:
        '200':
          description: Geri alındı
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileEntry' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: depsis_session
      description: |
        Sunucu tarafı oturum. JWT kullanılmaz — güvenlik olayında tüm oturumların iptal
        edilebilmesi gerekiyor ve stateless token bunu ucuz yapmıyor (ADR-0009).
  parameters:
    FileId:
      name: id
      in: path
      required: true
      description: Kararlı dosya kimliği. Yol asla kimlik olarak kullanılmaz (ADR-0005).
      schema: { type: string, format: uuid }
    Cursor:
      name: cursor
      in: query
      description: Opak. İstemci bunu üretmez ve ayrıştırmaz; kodlaması sunucuya aittir.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: |
        Kullanıcı ve uç kapsamında benzersiz. Aynı anahtar farklı bir istek parmak iziyle
        gelirse 409 döner. RFC yoktur — IETF taslağı Expired and archived durumundadır — bu
        yüzden semantik burada tanımlanır (ADR-0008).
      schema: { type: string, maxLength: 255 }
    IfMatch:
      name: If-Match
      in: header
      description: Çakışma kontrolü. Eşleşmezse 412.
      schema: { type: string }
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
  schemas:
...
    FileEntry:
      type: object
      required: [id, parentId, name, kind, size, modifiedAt, permissions]
      properties:
        permissions:
          type: array
          description: |
            ÇAĞIRANIN bu girdide yapabildikleri, miras hesaplandıktan sonra (ADR-0021).

            Satır başına, klasör başına değil, ve bunun bir maliyeti var: bir sayfa çizmek, o
            sayfadaki her girdi için ata zincirini yürümek demek. Yine de satır başına, çünkü
            alternatifi arayüzün klasörün iznine bakıp her satıra aynı düğmeleri koyması ve alt
            klasöre dar bir grant konmuş satırda o düğmenin 403 alması — yani sunucunun
            reddedeceği bir şeyi teklif eden bir arayüz.

            Sunucu bunu sayfa başına TEK ata yürüyüşüyle hesaplar: aynı klasördeki satırların
            atası ortak, ve yalnız kendi grant'ı olan satırlar ek iş çıkarır.
          items: { $ref: '#/components/schemas/FolderPermission' }
        id: { type: string, format: uuid }
        parentId:
          type: [string, 'null']
          format: uuid
        name: { type: string }
        kind:
          type: string
          enum: [file, folder]
        size: { type: integer, format: int64 }
        modifiedAt: { type: string, format: date-time }
        mimeType: { type: string }
        trashedAt:
          type: string
          format: date-time
          description: Yalnız çöp listesinde. Yoksa girdi çöpte değil.
        expiresAt:
          type: string
          format: date-time
          description: |
            Bu girdinin kalıcı olarak silineceği an: `trashedAt` + saklama süresi.

            Saklama politikası kapalıysa YOK, ve yokluğu "süresiz" demek. Bir tarih göstermek,
            sonra o tarihte hiçbir şey olmaması, çöp kutusunun tutmadığı bir söz olurdu.

            KENDİ ÜST DÜZEYİNDE olmayan bir girdi de yok: çöpe atılmış bir klasörün içindeki
            dosya, kendi tarihinde değil, kökünün tarihinde gider. Kendi tarihini göstermek,
            temizlemenin uymadığı bir geri sayım göstermek olurdu.
    FileEntryDetail:
      allOf:
        - $ref: '#/components/schemas/FileEntry'
        - type: object
          properties:
            createdAt: { type: string, format: date-time }
            ownerId: { type: string, format: uuid }
            checksum:
              type: string
              description: Sunucu tarafı kanonik digest.
            effectivePermissions:
              type: array
              items: { $ref: '#/components/schemas/Permission' }
    FileEntryPage:
      type: object
      required: [items, hasMore]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/FileEntry' }
        nextCursor: { type: string }
        hasMore:
          type: boolean
          description: |
            Toplam sayı bilerek yok. Filtrelenmemiş bir sayaç, kiracının göremediği
            satırların varlığını sızdırır.
...
    UpdateFileRequest:
      type: object
      minProperties: 1
      description: |
        Yeniden adlandırma ve taşıma tek uçta, çünkü ikisi de aynı satırın aynı iki kolonu:
        `name` ve `parent_id`. Dosya sistemi tarafında da tek bir rename(2) çağrısı.

        Taşıma AYNI PAYLAŞIM içinde. Paylaşımlar arası taşıma bir rename DEĞİL — ölçüldüğü üzere
        dataset'ler arası rename(2) EXDEV veriyor (ADR-0008) — ve kopyala+sil işine dönüşmesi
        gerekir; o iş /file-operations'ın konusu ve henüz yazılmadı. Başka bir paylaşımın
        klasörüne taşıma isteği 409 döner, sessizce yanlış yere koymaz.
      properties:
        name: { $ref: '#/components/schemas/FileName' }
        parentId:
          type: [string, 'null']
          format: uuid
          description: null, paylaşımın köküne taşır.
...
    JobRef:
      type: object
      required: [jobId]
      properties:
        jobId: { type: string, format: uuid }
...
    LinkTaskFileRequest:
      type: object
      required: [fileEntryId]
      properties:
        fileEntryId: { type: string, format: uuid }
...